test: isolate sandbox push service coverage
