+++ packages/control-plane/src/session/sandbox-events/processor.test.ts
@@ -592,192 +592,6 @@ describe("SessionSandboxEventProcessor", () => {
     );
   });
 
-  it("fails a push immediately when the command cannot be delivered", async () => {
-    vi.useFakeTimers();
-    try {
-      const h = createProcessor();
-      const sandboxWs = { readyState: WebSocket.OPEN } as WebSocket;
-      h.wsManager.getSandboxSocket.mockReturnValue(sandboxWs);
-      h.wsManager.send.mockReturnValue(false);
-
-      let result: Awaited<ReturnType<typeof h.pushService.pushBranchToRemote>> | undefined;
-      void h.pushService
-        .pushBranchToRemote(createPushSpec("acme", "web", "feature/test"))
-        .then((pushResult) => {
-          result = pushResult;
-        });
-      await vi.advanceTimersByTimeAsync(0);
-
-      expect(result).toEqual({
-        success: false,
-        error: expect.stringContaining("Failed to deliver push command to sandbox"),
-      });
-      expect(vi.getTimerCount()).toBe(0);
-
-      h.pushService.settlePush({
-        type: "push_complete",
-        branchName: "feature/test",
-        repoOwner: "acme",
-        repoName: "web",
-        timestamp: 1000,
-      });
-      expect(h.log.warn).toHaveBeenCalledWith(
-        "Push event matched no pending resolver",
-        expect.objectContaining({ pending_resolvers: [] })
-      );
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
-  describe("push resolver keying", () => {
-    function connectSandbox(h: ReturnType<typeof createProcessor>) {
-      const sandboxWs = { readyState: WebSocket.OPEN } as WebSocket;
-      h.wsManager.getSandboxSocket.mockReturnValue(sandboxWs);
-      return sandboxWs;
-    }
-
-    it("settles the matching push when two repos push the same branch name", async () => {
-      const h = createProcessor();
-      connectSandbox(h);
-
-      const webPush = h.pushService.pushBranchToRemote(
-        createPushSpec("acme", "web", "open-inspect/session-1")
-      );
-      const backendPush = h.pushService.pushBranchToRemote(
-        createPushSpec("acme", "backend", "open-inspect/session-1")
-      );
-
-      await h.processor.processSandboxEvent({
-        type: "push_error",
-        branchName: "open-inspect/session-1",
-        repoOwner: "acme",
-        repoName: "backend",
-        error: "remote rejected",
-        timestamp: 1000,
-      });
-      await h.processor.processSandboxEvent({
-        type: "push_complete",
-        branchName: "open-inspect/session-1",
-        repoOwner: "acme",
-        repoName: "web",
-        timestamp: 1001,
-      });
-
-      await expect(webPush).resolves.toEqual({ success: true });
-      await expect(backendPush).resolves.toEqual({
-        success: false,
-        error: expect.stringContaining("remote rejected"),
-      });
-    });
-
-    it("settles the sole pending push on a terminal event without repo identity", async () => {
-      const h = createProcessor();
-      connectSandbox(h);
-
-      const pushPromise = h.pushService.pushBranchToRemote(
-        createPushSpec("acme", "web", "feature/test")
-      );
-
-      // Legacy single-repo runtimes echo no repo identity.
-      await h.processor.processSandboxEvent({
-        type: "push_complete",
-        branchName: "feature/test",
-        timestamp: 1000,
-      });
-
-      await expect(pushPromise).resolves.toEqual({ success: true });
-    });
-
-    it("rejects the sole pending push on a branch-less push_error", async () => {
-      const h = createProcessor();
-      connectSandbox(h);
-
-      const pushPromise = h.pushService.pushBranchToRemote(
-        createPushSpec("acme", "web", "feature/test")
-      );
-
-      // The bridge's "no repository found" path emits push_error with no
-      // branchName at all; it must reject the pending push instead of
-      // leaking it to the 360 s timeout.
-      await h.processor.processSandboxEvent({
-        type: "push_error",
-        error: "No repository found for push",
-        timestamp: 1000,
-      });
-
-      await expect(pushPromise).resolves.toEqual({
-        success: false,
-        error: expect.stringContaining("No repository found for push"),
-      });
-    });
-
-    it("drops a fully identified event that mismatches the sole pending push", async () => {
-      const h = createProcessor();
-      connectSandbox(h);
-
-      const pushPromise = h.pushService.pushBranchToRemote(
-        createPushSpec("acme", "web", "feature/test")
-      );
-
-      // A stale event for a different repo must not settle the pending push
-      // just because it is the only one in flight.
-      await h.processor.processSandboxEvent({
-        type: "push_error",
-        branchName: "feature/test",
-        repoOwner: "acme",
-        repoName: "backend",
-        error: "remote rejected",
-        timestamp: 1000,
-      });
-
-      await h.processor.processSandboxEvent({
-        type: "push_complete",
-        branchName: "feature/test",
-        repoOwner: "acme",
-        repoName: "web",
-        timestamp: 1001,
-      });
-
-      await expect(pushPromise).resolves.toEqual({ success: true });
-    });
-
-    it("drops an identity-less terminal event when several pushes are pending", async () => {
-      const h = createProcessor();
-      connectSandbox(h);
-
-      const webPush = h.pushService.pushBranchToRemote(createPushSpec("acme", "web", "feature/a"));
-      const backendPush = h.pushService.pushBranchToRemote(
-        createPushSpec("acme", "backend", "feature/b")
-      );
-
-      await h.processor.processSandboxEvent({
-        type: "push_error",
-        error: "ambiguous",
-        timestamp: 1000,
-      });
-
-      // Neither push settles from the ambiguous event; identified events do.
-      await h.processor.processSandboxEvent({
-        type: "push_complete",
-        branchName: "feature/a",
-        repoOwner: "acme",
-        repoName: "web",
-        timestamp: 1001,
-      });
-      await h.processor.processSandboxEvent({
-        type: "push_complete",
-        branchName: "feature/b",
-        repoOwner: "acme",
-        repoName: "backend",
-        timestamp: 1002,
-      });
-
-      await expect(webPush).resolves.toEqual({ success: true });
-      await expect(backendPush).resolves.toEqual({ success: true });
-    });
-  });
-
   describe("activity tracking for intermediate events", () => {
     it("resets activity timer on tool_call", async () => {
       const h = createProcessor();
